Renames MeetupCard component to ExternalInfoCard.
