+++ components/ResourcesSection.vue
@@ -14,7 +14,7 @@
             <div
                 class="mt-8 rounded-lg bg-gray-200 overflow-hidden shadow divide-y divide-gray-200 sm:divide-y-0 sm:grid sm:grid-cols-2 sm:gap-px"
             >
-                <MeetupCard
+                <ExternalInfoCard
                     v-for="(resource, index) in resources"
                     :key="index"
                     :logo="resource.logo"
@@ -24,16 +24,16 @@
                     <p v-if="resource.meetingTime" class="mt-2">
                         {{ resource.meetingTime }}
                     </p>
-                </MeetupCard>
+                </ExternalInfoCard>
             </div>
         </div>
     </div>
 </template>
 
 <script>
-import MeetupCard from "./MeetupCard.vue";
+import ExternalInfoCard from "./ExternalInfoCard.vue";
 export default {
-    components: { MeetupCard },
+    components: { ExternalInfoCard },
     computed: {
         resources() {
             return this.$store.getters["resources/getAllResources"];
